aly can talk now i think*

+++ src/main/java/me/kambing/events/Troll.java
@@ -6,6 +6,8 @@ import net.dv8tion.jda.api.events.ReadyEvent;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 
+import static me.kambing.events.TrollToggler.troll;
+
 public class Troll extends ListenerAdapter {
 
     @Override
@@ -21,7 +23,7 @@ public class Troll extends ListenerAdapter {
         if (u == null)
             return;
 
-        if (m.getId().equals("")) {
+        if (m.getId().equals("823621414203686922") && troll) {
             u.delete().queue();
         }
     }
